Fixed Issue#17 by hooking into the Terminal Tab
Fixed issue #17 where action commands failed because everything was made upper before executing commands.
This fix removes the toUpperCase calls, and it uses the hooked Terminal Tab to execute the commands. The Terminal Tab logic handles cases when commands should NOT be made uppercase. It adheres to the Terminal Auto Uppercase Blacklist.
Used the following macro to test the change:
m300 s740 p25 ; Testing lowercase beep
; Testing line with only comments
; Testing action command with M118 starting in lower case
m118 // action:disconnect
Expected behavior:
Both of the above commands should work, and the comments should not get sent.
The Terminal Tab should handle the blacklist logic, and therefore any commands that are in the blacklist should only get the prefix command forced uppercase. The rest of the command should be sent as-is. All other command should be made fully uppercase.